Properly filter borrowings inside items
When providing a from_date or until_date, only the item.borrowings that
lie within the given range are returned as part of the borrowed-items/
api.

This was solved by explicitly writing the borrowings to a new property
called item.cached_borrowings while filtering. If we later call
item.get_borrowings, we simply return the cached property. A bit hacky,
but it works.