Replace raw sql by Django filtering
Up until now, borrowed items were fetched by using a raw sql statement
for efficiency reasons and because I didn't know of any way to do it
efficiently using Django.
This is very ugly however and makes filtering in the future difficult.
The currently implemented filtering should be almost as efficient as the
raw sql statement but allows us to do filtering, which is of course way
more flexible.